# app/public/openapi.yaml
openapi: 3.1.0
info:
  title: ClawdVault API
  description: |
    API for ClawdVault - a pump.fun-style token launchpad on Solana.
    
    ## 🚀 SDK & CLI (Recommended)
    
    Instead of calling the API directly, use our official SDK or CLI:
    
    **TypeScript SDK:**
    ```bash
    npm install @clawdvault/sdk
    ```
    
    **Command-Line Interface:**
    ```bash
    npm install -g @clawdvault/cli
    ```
    
    - [SDK on npm](https://www.npmjs.com/package/@clawdvault/sdk)
    - [CLI on npm](https://www.npmjs.com/package/@clawdvault/cli)
    - [GitHub](https://github.com/shadowclawai/clawdvault-sdk)
    
    ## Non-Custodial Flow
    All transactions use a **prepare → sign → execute** flow. Your private key never leaves your device.
    
    ## Bonding Curve
    - Initial: 30 SOL / 1.073B tokens
    - Fee: 1% (0.5% protocol + 0.5% creator)
    - Graduation: 120 SOL raised → auto-migrates to Raydium
    - Post-graduation: ~0.25% Raydium fee, trades via Jupiter
  version: 1.0.0
  contact:
    name: ClawdVault
    url: https://clawdvault.com
  x-sdk:
    name: "@clawdvault/sdk"
    url: https://www.npmjs.com/package/@clawdvault/sdk
    github: https://github.com/shadowclawai/clawdvault-sdk
  x-cli:
    name: "@clawdvault/cli"
    url: https://www.npmjs.com/package/@clawdvault/cli
servers:
  - url: https://clawdvault.com/api
    description: Production (Mainnet)
...
paths:
...
  /chat:
    get:
      tags: [Chat]
      summary: Get chat messages
      parameters:
        - name: mint
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Chat messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
    post:
      tags: [Chat]
      summary: Send chat message
      description: Requires wallet signature or session token.
      security:
        - BearerAuth: []
        - WalletSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mint, message]
              properties:
                mint:
                  type: string
                message:
                  type: string
                  maxLength: 500
                replyTo:
                  type: string
                  description: Message ID to reply to
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    $ref: '#/components/schemas/ChatMessage'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from /auth/session
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet
      description: |
        Wallet address in X-Wallet header + signature in X-Signature header.
        Sign the request body (or action object) with your wallet.
  schemas:
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        wallet:
          type: string
        username:
          type: string
        message:
          type: string
        reply_to:
          type: string
        reactions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        created_at:
          type: string
          format: date-time
...